Algorithm impl

+++ src/main/kotlin/com/github/mpps/fsrs/model/Card.kt
@@ -4,8 +4,8 @@ import java.time.OffsetDateTime
 
 data class Card(
     var due: OffsetDateTime,
-    var stability: Long,
-    var difficulty: Long,
+    var stability: Double,
+    var difficulty: Double,
     var elapsedDays: Long,
     var scheduledDays: Long,
     var reps: Long,
